Changed the way of tracking the number of visits
It was required to change the way of logging how many times each
directory is visited. Instead of logging every visit I decided to let
the script log each day the directory is visited. This way we can avoid
the effects of "visiting spikes" and provide a more realistic tracking